AccountManagerPlugin: Do the real check, and compatible to Trac 0.11 too, refs #6821 and #8663.
Plugin components can be activated implicitly, if the code is located in the
`plugins` directory of the enironment.  Method
`ComponentManager.is_component_enabled` returns `None` for such a case,
so it's not at all a reliable method to detect component activation.

Special thanks to David Byrne for testing and for critical hints on the issue.