Use calendar and task sorter as enabled in trac.ini.  Refs #9290.
Generally, this allows use of web admin to easily enable an implementation.

If none are enabled, this defaults to SimpleCalendar and SimpleSorter.

If one is enabled, it is used.

If more than one are enabled, the last one is used and a message is
logged.  (Note that ExtensionPoint() returns an unordered list so "last"
is rather arbitrary.)

This is a first step toward allowing user-provided calendars.