TracAnnouncer: Improve ticket change notification template, refs #8620.
Feels like there must be more corner-cases, that are not handled optimal now,
but this is an easy win, so won't miss it for older announcer versions as well.

Thanks to jc2trachacks for taking care to report as well as for providing a
good starting point towards the actual fix (omit the empty old value too).