Refine background ticket rescheduler.  Refs #9648.
Move the exclusion of milestone pseudo-tickets closer to where it makes
a difference (when we save the data).

This should be a no-op most of the time but without this the ticket set
sometimes has dangling references which makes the plugin fragile when
adding new function invocations in certain places.