1.0.0dev: Use long rev when creating `CodeReviewer` object. Refs #12412.
The `rev` string passed as a `CommitTicketReferenceMacro` arg will be shortened in the case of a hash (Git or Hg). The full hash is obtained after retrieving a `changeset` object. This change has only been tested with Subversion, but is hypothesized to fix an issue with the code review status in `CommitTicketReferenceMacro` not updating when using Git.