DoxygenPlugin, now version 0.4, proposes a new item in the Panel Admin, giving acces to a Web form. This form is automatically built from the Doxyfile given by the doxygen command running on the Trac serveur if any. The form is filled with the default values in the standard Doxyfile; the user may change them, afterwards he submits the form, the values are saved in a new Doxygen file and the Doxygen command is executed with this file as argument. On return, the  same Web form is shown, with the changes made before, but a message is first shown saying if the Doxygen command ran correctly or not. A click on the triangle in front of the message will show the wole trace of the command, or the error message if any.
The new Doxygen file is kept on the server, so this Web page, once filled with the proper values, can be seen as a way to execute the Doxygen command on the Trac server.

This way of combining Trac and Doxygen on a unique server is different from the demand in the ticket system, but we may consider it closes #4740.